removed binary test collection
